+++ src/client/debugger/constants.ts
@@ -4,7 +4,7 @@
 'use strict';
 
 import * as path from 'path';
-import { EXTENSION_ROOT_DIR } from '../../common/constants';
+import { EXTENSION_ROOT_DIR } from '../common/constants';
 
 export const PTVSD_PATH = path.join(EXTENSION_ROOT_DIR, 'pythonFiles', 'experimental', 'ptvsd');
 export const DebuggerTypeName = 'python';
