+++ src/client/debugger/debugAdapter/Common/protocolWriter.ts
