+++ src/client/debugger/debugAdapter/Common/telemetry.ts
@@ -4,9 +4,9 @@
 // tslint:disable:no-function-expression no-any no-invalid-this no-use-before-declare
 
 import { DebugSession, StoppedEvent } from 'vscode-debugadapter';
-import { StopWatch } from '../../common/utils/stopWatch';
-import { DEBUGGER_PERFORMANCE } from '../../telemetry/constants';
-import { DebuggerPerformanceTelemetry } from '../../telemetry/types';
+import { StopWatch } from '../../../common/utils/stopWatch';
+import { DEBUGGER_PERFORMANCE } from '../../../telemetry/constants';
+import { DebuggerPerformanceTelemetry } from '../../../telemetry/types';
 import { TelemetryEvent } from './Contracts';
 
 type DebugAction = 'stepIn' | 'stepOut' | 'continue' | 'next' | 'launch';
