+++ src/client/debugger/debugAdapter/DebugClients/LocalDebugClient.ts
@@ -2,13 +2,14 @@ import { ChildProcess, spawn } from 'child_process';
 import * as path from 'path';
 import { DebugSession, OutputEvent } from 'vscode-debugadapter';
 import { DebugProtocol } from 'vscode-debugprotocol';
-import { open } from '../../common/open';
-import { PathUtils } from '../../common/platform/pathUtils';
-import { CurrentProcess } from '../../common/process/currentProcess';
-import { noop } from '../../common/utils/misc';
-import { EnvironmentVariablesService } from '../../common/variables/environment';
-import { IServiceContainer } from '../../ioc/types';
-import { DebugOptions, IDebugServer, LaunchRequestArguments } from '../Common/Contracts';
+import { open } from '../../../common/open';
+import { PathUtils } from '../../../common/platform/pathUtils';
+import { CurrentProcess } from '../../../common/process/currentProcess';
+import { noop } from '../../../common/utils/misc';
+import { EnvironmentVariablesService } from '../../../common/variables/environment';
+import { IServiceContainer } from '../../../ioc/types';
+import { DebugOptions, LaunchRequestArguments } from '../../types';
+import { IDebugServer } from '../Common/Contracts';
 import { IS_WINDOWS } from '../Common/Utils';
 import { BaseDebugServer } from '../DebugServers/BaseDebugServer';
 import { LocalDebugServerV2 } from '../DebugServers/LocalDebugServerV2';
