+++ src/client/debugger/debugAdapter/DebugClients/nonDebugClientV2.ts
@@ -5,7 +5,7 @@
 
 import { ChildProcess } from 'child_process';
 import { DebugSession } from 'vscode-debugadapter';
-import { LaunchRequestArguments } from '../Common/Contracts';
+import { LaunchRequestArguments } from '../../types';
 import { IDebugLauncherScriptProvider } from '../types';
 import { DebugType } from './DebugClient';
 import { LocalDebugClientV2 } from './localDebugClientV2';
