+++ src/client/debugger/extension/types.ts
@@ -1,0 +1,10 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+'use strict';
+
+export const IDebugConfigurationProvider = Symbol('DebugConfigurationProvider');
+export const IDebuggerBanner = Symbol('IDebuggerBanner');
+export interface IDebuggerBanner {
+    initialize(): void;
+}
